requirements/base.txt
Django==4.2.4
psycopg2==2.9.7
django-environ
requests 2.31.0